Allows selecting text and clicking on links in all tooltips
